extract mock vault resolver to its own file
